Refactor static market snapshots into shared config
